requirements.txt
allure-pytest==2.9.45
selenium==3.141.0
pytest==6.2.5
pytest-xdist==2.5.0
requests==2.27.1